// src/pages/types.ts
export interface DonationRequestFields {
  male_clothes_items: string[];
  telegram_nickname: string[]
  phone_number: string[]
  preferred_time: string[]
  household_items: string[]
  female_clothes_items: string[]
  phone: string[]
  male_clothes_table_sizes: string[]
  children_clothes_items: string[]
  preferred_delivery_date: string[]
  name: string[]
  male_clothes_params: string[]
  children_shoe_sizes: string[]
  timestamp: string[]
  female_clothes_params: string[]
  tech_items: string[]
  nickname: string[]
  item_list: string[]
  household_tech_items: string[]
  comments: string[]
  male_shoe_sizes: string[]
  extra_items: string[]
  height: string[]
  female_clothes_table_sizes: string[]
  item_volume: string[]
  female_shoe_sizes: string[]
  kids_items: string[]
}